--- lib/pharos/local_client.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module Pharos
-  class LocalClient < Pharos::Transport
-    def connect
-      nil
-    end
-
-    def connected?
-      true
-    end
-
-    def interactive_session
-      return unless ENV['SHELL']
-      synchronize { system ENV['SHELL'] }
-    end
-
-    private
-
-    def command(cmd, **options)
-      LocalCommand.new(self, cmd, **options)
-    end
-  end
-end
