+++ lib/pharos/transport/local.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+module Pharos
+  module Transport
+    class Local < Base
+      def connect
+        nil
+      end
+
+      def connected?
+        true
+      end
+
+      def interactive_session
+        return unless ENV['SHELL']
+        synchronize { system ENV['SHELL'] }
+      end
+
+      private
+
+      def command(cmd, **options)
+        LocalCommand.new(self, cmd, **options)
+      end
+    end
+  end
+end
